# ✅ feat: Complete Phase E.3 Accessibility File Splitting for MCP Tool Compatibility (#84)