--- a/.kiro/specs/accessibility-file-splitting-phase-e3/tasks.md
+++ b/.kiro/specs/accessibility-file-splitting-phase-e3/tasks.md
@@ -1,6 +1,6 @@
 # Implementation Plan
 
-- [ ] 1. Project setup and accessibility analysis
+- [x] 1. Project setup and accessibility analysis
   - Create directory structures for all accessibility split components
   - Set up file size monitoring for target accessibility files
   - Analyze current test coverage for all 6 target accessibility files