--- a/.kiro/specs/accessibility-file-splitting-phase-e3/tasks.md
+++ b/.kiro/specs/accessibility-file-splitting-phase-e3/tasks.md
@@ -9,28 +9,28 @@
   - _Requirements: 1.1, 8.1, 8.3, 9.1_
 
 - [ ] 2. KeyboardNavigationTester.js splitting (3,116 words → 4 components)
-  - [ ] 2.1 Create KeyboardEventHandler component
+  - [x] 2.1 Create KeyboardEventHandler component
     - Extract keyboard event processing methods (`setupEventListeners`, `simulateKeyPress`, `detectKeyboardEvents`)
     - Implement event listener management and key combination detection
     - Create unit tests for keyboard event handling functionality
     - Ensure keyboard event compatibility with assistive technologies
     - _Requirements: 2.1, 2.2, 9.1, 11.5_
 
-  - [ ] 2.2 Create NavigationStateManager component
+  - [x] 2.2 Create NavigationStateManager component
     - Extract focus management methods (`trackFocusChanges`, `validateTabOrder`, `testFocusContainment`)
     - Implement tab order calculation and focus state tracking
     - Create unit tests for navigation state management functionality
     - Verify focus management maintains WCAG compliance
     - _Requirements: 2.1, 2.2, 9.1, 11.1_
 
-  - [ ] 2.3 Create KeyboardAccessibilityReporter component
+  - [x] 2.3 Create KeyboardAccessibilityReporter component
     - Extract reporting methods (`generateReport`, `categorizeIssues`, `calculateAccessibilityScore`)
     - Implement accessibility compliance scoring and issue categorization
     - Create unit tests for accessibility reporting functionality
     - Ensure reports maintain WCAG guideline references
     - _Requirements: 2.1, 2.2, 9.1, 11.1_
 
-  - [ ] 2.4 Refactor main KeyboardNavigationTester to use sub-components
+  - [x] 2.4 Refactor main KeyboardNavigationTester to use sub-components
     - Update constructor to initialize sub-components with accessibility context
     - Modify public methods to delegate to appropriate sub-components
     - Maintain existing public API for backward compatibility