--- a/.kiro/specs/accessibility-file-splitting-phase-e3/tasks.md
+++ b/.kiro/specs/accessibility-file-splitting-phase-e3/tasks.md
@@ -8,7 +8,7 @@
   - Verify WCAG 2.1 AA compliance baseline before splitting
   - _Requirements: 1.1, 8.1, 8.3, 9.1_
 
-- [ ] 2. KeyboardNavigationTester.js splitting (3,116 words → 4 components)
+- [x] 2. KeyboardNavigationTester.js splitting (3,116 words → 4 components)
   - [x] 2.1 Create KeyboardEventHandler component
     - Extract keyboard event processing methods (`setupEventListeners`, `simulateKeyPress`, `detectKeyboardEvents`)
     - Implement event listener management and key combination detection
@@ -37,7 +37,7 @@
     - Implement accessibility-aware error handling and fallback mechanisms
     - _Requirements: 2.3, 2.4, 2.5, 10.1, 10.2_
 
-  - [ ] 2.5 Verify KeyboardNavigationTester split completion
+  - [x] 2.5 Verify KeyboardNavigationTester split completion
     - Check all files are under 2,500 words
     - Run existing tests to ensure keyboard navigation functionality works
     - Test focus management and keyboard trap detection