--- a/.kiro/specs/accessibility-file-splitting-phase-e3/tasks.md
+++ b/.kiro/specs/accessibility-file-splitting-phase-e3/tasks.md
@@ -44,36 +44,36 @@
     - Verify WCAG 2.1 AA compliance is maintained
     - _Requirements: 1.1, 2.5, 9.2, 11.1_
 
-- [ ] 3. WCAGValidator.js splitting (2,931 words → 4 components)
-  - [ ] 3.1 Create WCAGRuleEngine component
+- [x] 3. WCAGValidator.js splitting (2,931 words → 4 components)
+  - [x] 3.1 Create WCAGRuleEngine component
     - Extract WCAG test methods (`runTest`, `testAltText`, `testColorContrast`, `testKeyboardNavigation`)
     - Implement guideline definitions and rule-specific validation logic
     - Create unit tests for WCAG rule engine functionality
     - Ensure all WCAG 2.1 AA guidelines are properly implemented
     - _Requirements: 3.1, 3.2, 3.5, 9.1_
 
-  - [ ] 3.2 Create AccessibilityAuditor component
+  - [x] 3.2 Create AccessibilityAuditor component
     - Extract auditing methods (`validateCategory`, `runQuickValidation`, `auditAccessibility`)
     - Implement category-based validation and issue classification
     - Create unit tests for accessibility auditing functionality
     - Verify comprehensive WCAG compliance checking
     - _Requirements: 3.1, 3.2, 3.5, 9.1_
 
-  - [ ] 3.3 Create ComplianceReporter component
+  - [x] 3.3 Create ComplianceReporter component
     - Extract reporting methods (`calculateOverallScore`, `updateTrends`, `saveValidationResults`)
     - Implement compliance scoring and trend analysis logic
     - Create unit tests for compliance reporting functionality
     - Ensure accurate WCAG compliance scoring
     - _Requirements: 3.1, 3.2, 3.5, 9.1_
 
-  - [ ] 3.4 Refactor main WCAGValidator to use sub-components
+  - [x] 3.4 Refactor main WCAGValidator to use sub-components
     - Update constructor to initialize sub-components with WCAG context
     - Modify public methods to delegate to appropriate sub-components
     - Maintain existing public API for backward compatibility
     - Implement WCAG-compliant error handling and fallback mechanisms
     - _Requirements: 3.3, 3.4, 3.5, 10.1, 10.2_
 
-  - [ ] 3.5 Verify WCAGValidator split completion
+  - [x] 3.5 Verify WCAGValidator split completion
     - Check all files are under 2,500 words
     - Run existing tests to ensure WCAG validation functionality works
     - Test all WCAG 2.1 AA guideline validation