… .kiro/specs/accessibility-file-splitting-phase-e3/design.md
… .kiro/specs/accessibility-file-splitting-phase-e3/requirements.md
… CLAUDE.md
… docs/accessibility-api-documentation.md
… docs/accessibility-deployment-strategy.md
… docs/accessibility-migration-guide.md
… docs/accessibility-phase-e3-final-report.md
… docs/accessibility-testing-procedures.md
… docs/architecture.md
… scripts/integration-testing.js
… scripts/keyboard-navigation-test.js
… scripts/performance-impact-assessment.js
… scripts/screen-reader-compatibility-test.js
… src/accessibility/AccessibilityOnboarding.js
… src/accessibility/AccessibilitySettingsUI.js
… src/accessibility/ColorContrastAnalyzer.js
… src/accessibility/KeyboardNavigationTester.js
… src/accessibility/ScreenReaderSimulator.js
… src/accessibility/WCAGValidator.js
… src/accessibility/color-contrast/ColorAnalysisEngine.js
… src/accessibility/color-contrast/ColorBlindnessSimulator.js
… src/accessibility/color-contrast/ContrastCalculator.js
… src/accessibility/keyboard-navigation/KeyboardAccessibilityReporter.js
… src/accessibility/keyboard-navigation/KeyboardEventHandler.js
… src/accessibility/keyboard-navigation/NavigationStateManager.js
… src/accessibility/onboarding/OnboardingFlowManager.js
… src/accessibility/onboarding/OnboardingProgressTracker.js
… src/accessibility/screen-reader/ARIAAttributeProcessor.js
… src/accessibility/screen-reader/ScreenReaderEngine.js
… src/accessibility/screen-reader/TextToSpeechController.js